+++ toad/src/ttable.rs
@@ -78,7 +78,7 @@ impl TTableEntry {
         score: Score,
         bounds: SearchBounds,
         depth: u8,
-        ply: i32,
+        ply: i8,
     ) -> Self {
         // Determine what kind of node this is fist, before score adjustment
         let node_type = NodeType::new(score, bounds);
@@ -102,7 +102,7 @@ impl TTableEntry {
     ///     2. The entry is an upper bound ([`NodeType::All`]) and its score is `<= alpha`.
     ///     3. The entry is a lower bound ([`NodeType::Cut`]) and its score is `>= beta`.
     #[inline(always)]
-    pub fn try_score(&self, bounds: SearchBounds, ply: i32) -> Option<Score> {
+    pub fn try_score(&self, bounds: SearchBounds, ply: i8) -> Option<Score> {
         // Adjust mate scores to be relative to current ply
         let score = if self.score.is_mate() {
             self.score.relative(ply)
